Revert Sobol64 distributions (#289)
* restore old sobol64 uniform distribution, minor fixes

* update changelog, disable python test

[ROCm/rocRAND commit: 75467b4a102d7fce2de1a103bc0de1d58beb095a]

+++ library/include/rocrand/rocrand_log_normal.h
@@ -728,7 +728,7 @@ FQUALIFIERS
 float rocrand_log_normal(rocrand_state_sobol64* state, float mean, float stddev)
 {
     float r = rocrand_device::detail::normal_distribution(rocrand(state));
-    return exp(mean + (stddev * r));
+    return expf(mean + (stddev * r));
 }
 
 /**
@@ -766,7 +766,7 @@ FQUALIFIERS
 float rocrand_log_normal(rocrand_state_scrambled_sobol64* state, float mean, float stddev)
 {
     float r = rocrand_device::detail::normal_distribution(rocrand(state));
-    return exp(mean + (stddev * r));
+    return expf(mean + (stddev * r));
 }
 
 /**
